CI: Install `wheel` and use `--use-pep517` to `pip install`
The `pyscf` package does still use the legacy `setup.py` for the install
which leads to an error in the `pip install` for Python 3.11. Since
there are also no wheels available, the only solution is to build the
wheel ourselves using `wheel` and by passing `--use-pep517` option to
`pip`. The build of the wheel can take quite a bit of time so the
timeout of the CI workflow is increased significantly.