`PyscfCalculation`: Improve templating of the input script
The `script.py` input script is generated through the Jinja2 templating
engine. The `script.py.j2` is the main template that includes sub
templates. The original implementation included the sub templates as
literal strings that would be rendered separately, with the rendered sub
template string being passed as a variable to the main template.

This use-case is directly provided by Jinja2 through the `include`
filter. This prevents us from having to render the sub templates
manually and simply pass in all variables to the main template and
render it once. This significantly simplified the code.