Add the `CudftCalculation` calculation job plugin
This plugin is a subclass of the `PyscfCalculation` and enables running
CuDFT through AiiDA. The main difference for running CuDFT compared to
PySCF is that the former requires a few additional lines in the input
script. This is accomplished by overriding the `render_script` method
and rendering a custom script template `script_cudft.py.j2` which
includes these required lines.